Rename dependency injection module
